Refactor: Split modules into service-specific modules and created `user-service` and `login-service`
- Separated `core-shared`, `core-network-resources`, and `core-network-model` into dedicated modules for login and user services.
  - New modules: `login-service-network-resources`, `login-service-network-model`, `users-service-network-resources`, `users-service-network-model`.
- Renamed and moved relevant classes and files to their new respective service-specific modules.
  - `AlmacenLoginRequestResource` -> `LoginRequestResource` (in login service)
  - `AlmacenUserResource` -> `UserResource` (in user service)
  - `AlmacenUserModel` -> `UserModel` (in user service)
  - `AlmacenUserPasswordChangeModel` -> `UserPasswordChangeModel` (in user service)
  - `AlmacenLoginRequestModel` -> `LoginRequestModel` (in login service)
  - `AlmacenLoginResponseModel` -> `LoginResponseModel` (in login service)
  - `RefreshTokenRequest` -> `RefreshRequestModel` (in login service)
  - `UserRoleModel` moved from `almacen-service-network-model` to `users-service-network-model`.
- Removed `ResourceRoute` interface and updated resource classes to no longer implement it.
  - `getRoute()` methods are now directly defined within the resource classes.
- Updated `build.gradle.kts` files:
  - Changed group IDs for `almacen-service-network-resources` and `almacen-service-network-model` to `com.cocot3ro.gh.deps`.
  - Removed `coreShared` dependency from `almacen-service-network-resources`.
  - Added KSP plugin to `login-service-network-model` and `users-service-network-model`.
- Updated `settings.gradle.kts` to include the new modules and remove the old core modules.
