--- almacen-service-network-model/src/main/kotlin/com/cocot3ro/gh/almacen/data/network/model/UserRoleModel.kt
@@ -1,6 +1,0 @@
-package com.cocot3ro.gh.almacen.data.network.model
-
-enum class UserRoleModel {
-    ADMIN,
-    USER
-}
