+++ almacen-service-network-resources/src/main/kotlin/com/cocot3ro/gh/almacen/data/network/resources/AlmacenResource.kt
@@ -1,27 +1,26 @@
 package com.cocot3ro.gh.almacen.data.network.resources
 
-import com.cocot3ro.gh.core.ResourceRoute
 import io.ktor.resources.Resource
 
 @Resource(path = AlmacenResource.PATH)
-class AlmacenResource : ResourceRoute {
+class AlmacenResource {
 
     companion object {
         const val PATH: String = "/api/almacen"
     }
 
-    override fun getRoute(): String = PATH
+    fun getRoute(): String = PATH
 
     @Resource(path = Status.PATH)
     data class Status(
-        val parent: AlmacenResource = AlmacenResource(),
-    ) : ResourceRoute {
+        val parent: AlmacenResource = AlmacenResource()
+    ) {
 
         companion object {
             const val PATH: String = "status"
         }
 
-        override fun getRoute(): String = "${parent.getRoute()}/$PATH"
+        fun getRoute(): String = "${parent.getRoute()}/$PATH"
     }
 
 }
