--- core-network-model/src/main/kotlin/com/cocot3ro/gh/core/RefreshTokenRequest.kt
@@ -1,8 +1,0 @@
-package com.cocot3ro.gh.core
-
-import kotlinx.serialization.Serializable
-
-@Serializable
-data class RefreshTokenRequest(
-    val refreshToken: String
-)
