--- core-network-resources/build.gradle.kts
@@ -1,50 +1,0 @@
-import org.jetbrains.kotlin.gradle.dsl.JvmTarget
-
-plugins {
-    id("java-library")
-    alias(libs.plugins.jetbrainsKotlinJvm)
-    `maven-publish`
-
-    alias(libs.plugins.kotlin.plugin.serialization)
-}
-
-group = "com.cocot3ro.gh"
-version = "1.0.0-SNAPSHOT"
-
-java {
-    sourceCompatibility = JavaVersion.VERSION_11
-    targetCompatibility = JavaVersion.VERSION_11
-}
-
-kotlin {
-    compilerOptions {
-        jvmTarget = JvmTarget.JVM_11
-    }
-}
-
-dependencies {
-    implementation(projects.coreShared)
-
-    implementation(platform(libs.ktor.bom))
-    implementation(libs.ktor.server.resources)
-    implementation(libs.ktor.serialization.kotlinx.json)
-}
-
-publishing {
-    publications {
-        create<MavenPublication>("gpr") {
-            from(components["java"])
-        }
-    }
-
-    repositories {
-        maven {
-            name = "GithubPackages"
-            url = uri("https://maven.pkg.github.com/cocot3ro/GestionHotel-deps")
-            credentials {
-                username = project.findProperty("gpr.user") as String
-                password = project.findProperty("gpr.key") as String
-            }
-        }
-    }
-}
