--- core-shared/src/main/java/com/cocot3ro/gh/core/ResourceRoute.kt
@@ -1,7 +1,0 @@
-package com.cocot3ro.gh.core
-
-interface ResourceRoute {
-
-    fun getRoute(): String
-
-}
