+++ login-service-network-model/src/main/kotlin/com/cocot3ro/gh/services/login/LoginRequestModel.kt
@@ -1,0 +1,12 @@
+package com.cocot3ro.gh.services.login
+
+import kotlinx.serialization.SerialName
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class LoginRequestModel(
+    @SerialName("user_id")
+    val userId: Long,
+
+    val password: String?
+)
