+++ login-service-network-model/src/main/kotlin/com/cocot3ro/gh/services/login/RefreshRequestModel.kt
@@ -1,0 +1,8 @@
+package com.cocot3ro.gh.services.login
+
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class RefreshRequestModel(
+    val refreshToken: String
+)
