+++ login-service-network-resources/.gitignore
