+++ login-service-network-resources/build.gradle.kts
