+++ users-service-network-model/src/main/kotlin/com/cocot3ro/gh/services/users/UserRoleModel.kt
@@ -1,0 +1,6 @@
+package com.cocot3ro.gh.services.users
+
+enum class UserRoleModel {
+    ADMIN,
+    USER
+}
