+++ users-service-network-resources/build.gradle.kts
